# deployment/Swagger/v1.json
{"swagger":"2.0","info":{"version":"v1","title":"hms.entappsettings.webapi"},"host":"localhost:44319","schemes":["https"],"paths":{"/api/AppSettingGroups/{appSettingGroupId}":{"get":{"tags":["AppSettingGroups"],"summary":"Returns the App Setting Groups details for a single group Id","operationId":"AppSettingGroups_GetAppSettingGroup","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"appSettingGroupId","in":"path","description":"","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AppSettingGroupDTO"}}}},"delete":{"tags":["AppSettingGroups"],"summary":"RESTRICTED: Delete an App Setting Group","operationId":"AppSettingGroups_DeleteAppSettingGroup","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"appSettingGroupId","in":"path","description":"","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AppSettingGroupDTO"}},"404":{"description":"Not Found"}}}},"/api/AppSettingGroups/ByPath":{"get":{"tags":["AppSettingGroups"],"summary":"Returns the App Setting Groups details for a given Group Path","operationId":"AppSettingGroups_GetAppSettingGroupByPath","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"groupPath","in":"query","description":"The groups path in the format '\\Core\\Grandparent\\Parent\\Child'. MUST be URL encoded e.g. %5CCore%5CGrandParent%5CParent","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AppSettingGroupDTO"}},"400":{"description":"Bad Request: Null or missing parameter"}}}},"/api/AppSettingGroups/{id}":{"put":{"tags":["AppSettingGroups"],"summary":"RESTRICTED: Update an existing App Setting Group","operationId":"AppSettingGroups_PutAppSettingGroup","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":[],"parameters":[{"name":"id","in":"path","description":"","required":true,"type":"integer","format":"int32"},{"name":"appSettingGroupDTO","in":"body","description":"","required":true,"schema":{"$ref":"#/definitions/AppSettingGroupDTO"}}],"responses":{"204":{"description":"Success, No Content"},"400":{"description":"Bad Request: Id does not match the body"},"404":{"description":"Not Found: AppSettingGroup Id not found"}}}},"/api/AppSettingGroups":{"get":{"tags":["AppSettingGroups"],"summary":"Returns all the App Setting Groups","operationId":"AppSettingGroups_GetAppSettingGroups","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/AppSettingGroupDTO"}}}}},"post":{"tags":["AppSettingGroups"],"summary":"RESTRICTED: Add an App Setting Group","operationId":"AppSettingGroups_PostAppSettingGroup","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"appSettingGroupDTO","in":"body","description":"","required":true,"schema":{"$ref":"#/definitions/AppSettingGroupDTO"}}],"responses":{"201":{"description":"Success, Created. See location header value for URI to newly created resource.","schema":{"$ref":"#/definitions/AppSettingGroupDTO"}},"400":{"description":"Bad Request: Passed model invalid"},"409":{"description":"Conflict: Could not save due to a conflict"}}}},"/api/AppSettings/{tenantId}/{appSettingGroupId}":{"get":{"tags":["AppSettings"],"summary":"Returns RESULTANT App Settings for a given Tenant and AppSettingsGroup. NO internal settings published.","operationId":"AppSettings_GetAppSettings","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"tenantId","in":"path","description":"The Id of the Tenant whose settings to return. This should be hard configured in the internal application config.","required":true,"type":"integer","format":"int32"},{"name":"appSettingGroupId","in":"path","description":"The AppSetting Group Id (i.e. the type of client). This should be hard configured in the internal application config.","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/AppSettingDTO"}}},"404":{"description":"Not Found"}}}},"/api/AppSettings/WithOverrideFlag/{tenantId}/{appSettingGroupId}":{"get":{"tags":["AppSettings"],"summary":"RESTRICTED: Returns App Settings for a given Tenant and App Setting Group with an Overridden flag.","operationId":"AppSettings_GetAppSettingsWithOverride","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"tenantId","in":"path","description":"The Id of the Tenant whose settings to return. This should be hard configured in the internal application config.","required":true,"type":"integer","format":"int32"},{"name":"appSettingGroupId","in":"path","description":"The AppSetting Group Id (i.e. the type of client). This should be hard configured in the internal application config.","required":true,"type":"integer","format":"int32"},{"name":"includeInternals","in":"query","description":"Include internal settings. Default is false","required":false,"type":"boolean"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/AppSettingWithOverrideDTO"}}}}}},"/api/AppSettings/{appSettingId}":{"put":{"tags":["AppSettings"],"summary":"RESTRICTED: Update an existing Tenant","operationId":"AppSettings_PutAppSetting","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":[],"parameters":[{"name":"appSettingId","in":"path","description":"","required":true,"type":"integer","format":"int32"},{"name":"appSettingDTO","in":"body","description":"","required":true,"schema":{"$ref":"#/definitions/AppSettingDTO"}}],"responses":{"204":{"description":"OK. No content"},"400":{"description":"Bad Request: Id does not match the body"},"404":{"description":"Not Found: Tenant Id not found"}}},"delete":{"tags":["AppSettings"],"summary":"RESTRICTED: Delete a Tenant","operationId":"AppSettings_DeleteTenant","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"appSettingId","in":"path","description":"The App Setting Id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AppSettingDTO"}},"404":{"description":"Not Found"}}}},"/api/AppSettings":{"get":{"tags":["AppSettings"],"summary":"RESTRICTED: Gets all App Settings","operationId":"AppSettings_GetAllAppSettings","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"includeInternals","in":"query","description":"Include internal falgged settings. Default is false","required":false,"type":"boolean"},{"name":"skip","in":"query","description":"default = 0","required":false,"type":"integer","format":"int32"},{"name":"take","in":"query","description":"default = 20","required":false,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/AppSettingDTO"}}},"404":{"description":"Not Found"}}},"post":{"tags":["AppSettings"],"summary":"RESTRICTED: Adds an App Setting","operationId":"AppSettings_PostAppSetting","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"appSettingDTO","in":"body","description":"A new App Setting","required":true,"schema":{"$ref":"#/definitions/AppSettingDTO"}}],"responses":{"201":{"description":"Success, Created. See location header value for URI to newly created resource.","schema":{"$ref":"#/definitions/AppSettingDTO"}},"400":{"description":"Bad Request: Passed model invalid"},"409":{"description":"Conflict: Could not save due to a conflict"}}}},"/api/AppSettingSections/{appSettingSectionId}":{"get":{"tags":["AppSettingSections"],"summary":"Get App Setting Section details by id","operationId":"AppSettingSections_GetAppSettingSection","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"appSettingSectionId","in":"path","description":"App Setting Section Id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AppSettingSectionDTO"}}}},"put":{"tags":["AppSettingSections"],"summary":"Update an existing App Setting Section","operationId":"AppSettingSections_PutAppSettingSection","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":[],"parameters":[{"name":"appSettingSectionId","in":"path","description":"App Setting Section Id","required":true,"type":"integer","format":"int32"},{"name":"appSettingSectionDTO","in":"body","description":"App Setting Section DTO","required":true,"schema":{"$ref":"#/definitions/AppSettingSectionDTO"}}],"responses":{"204":{"description":"OK. No content"},"400":{"description":"Bad Request: Id does not match the body"},"404":{"description":"Not Found: AppSettingSection Id not found"}}},"delete":{"tags":["AppSettingSections"],"summary":"Delete a AppSettingSection","operationId":"AppSettingSections_DeleteAppSettingSection","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"appSettingSectionId","in":"path","description":"The App Setting Section Id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AppSettingSectionDTO"}},"404":{"description":"Not Found"}}}},"/api/AppSettingSections":{"get":{"tags":["AppSettingSections"],"summary":"Returns all Sections","operationId":"AppSettingSections_GetAppSettingSections","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/AppSettingSectionDTO"}}},"500":{"description":"Internal Server Error"}}},"post":{"tags":["AppSettingSections"],"summary":"Add an App Setting Section","operationId":"AppSettingSections_PostAppSettingSection","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"appSettingSectionDTO","in":"body","description":"The new appSettingSection DTO","required":true,"schema":{"$ref":"#/definitions/AppSettingSectionDTO"}}],"responses":{"201":{"description":"Success, Created. See location header value for URI to newly created resource.","schema":{"$ref":"#/definitions/AppSettingSectionDTO"}},"400":{"description":"Bad Request: Passed model invalid"}}}},"/api/Tenants/{id}":{"get":{"tags":["Tenants"],"summary":"Get tenant details by tenantId","operationId":"Tenants_GetTenant","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"id","in":"path","description":"","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TenantDTO"}}}},"delete":{"tags":["Tenants"],"summary":"Delete a Tenant","operationId":"Tenants_DeleteTenant","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"id","in":"path","description":"The TenantId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TenantDTO"}},"404":{"description":"Not Found"}}}},"/api/Tenants/{tenantId}":{"put":{"tags":["Tenants"],"summary":"Update an existing Tenant","operationId":"Tenants_PutTenant","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":[],"parameters":[{"name":"tenantId","in":"path","description":"","required":true,"type":"integer","format":"int32"},{"name":"tenantDTO","in":"body","description":"","required":true,"schema":{"$ref":"#/definitions/TenantDTO"}}],"responses":{"204":{"description":"OK. No content"},"400":{"description":"Bad Request: Id does not match the body"},"404":{"description":"Not Found: Tenant Id not found"}}}},"/api/Tenants":{"get":{"tags":["Tenants"],"summary":"Returns all Tenants","operationId":"Tenants_GetTenants","consumes":[],"produces":["application/json","text/json","application/xml","text/xml"],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}},"post":{"tags":["Tenants"],"summary":"Add an Tenant","operationId":"Tenants_PostTenant","consumes":["application/json","text/json","application/xml","text/xml","application/x-www-form-urlencoded"],"produces":["application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"tenantDTO","in":"body","description":"The new tenant DTO","required":true,"schema":{"$ref":"#/definitions/TenantDTO"}}],"responses":{"201":{"description":"Success, Created. See location header value for URI to newly created resource.","schema":{"$ref":"#/definitions/TenantDTO"}},"400":{"description":"Bad Request: Passed model invalid"},"409":{"description":"Conflict: Could not save due to a conflict"}}}}},"definitions":{"AppSettingGroupDTO":{"description":"App Setting Group","type":"object","properties":{"AppSettingGroupId":{"format":"int32","description":"Unique App Setting Group Id","type":"integer"},"ParentGroupId":{"format":"int32","description":"Parent Group Id","type":"integer"},"Group":{"description":"Group name","type":"string"},"Description":{"description":"Group Description","type":"string"},"GroupPath":{"description":"Group Path","type":"string"},"ModifiedDate":{"format":"date-time","description":"Date last modified","type":"string"},"ModifiedBy":{"description":"Last modified by","type":"string"}}},"AppSettingDTO":{"description":"Application Setting","required":["SettingKey","TenantId","SettingGroupId","SettingSectionId","TypeId"],"type":"object","properties":{"AppSettingId":{"format":"int32","description":"The setting Id","type":"integer"},"SettingKey":{"description":"Setting Key","maxLength":50,"minLength":0,"type":"string"},"TenantId":{"format":"int32","description":"Tenant Id. Zero indicates that this setting applies to all affiliates unless overridden","type":"integer"},"SettingGroupId":{"description":"The group that this setting applies to. Each client is will be interested in the settings contained in a group and it's parent groups","type":"string"},"GroupPath":{"description":"The group path","type":"string"},"SettingSectionId":{"description":"The setting UI section","type":"string"},"TypeId":{"description":"The type of setting value. e.g. HTML, TEXT,","enum":["TEXT","INTEGER","DECIMAL","BOOL","CSV","HTML","JSON","XML"],"type":"string"},"SettingValue":{"description":"The setting value.","type":"string"},"IsLocked":{"description":"Indicates that this setting can not be edited in the DCR Admin UI","type":"boolean"},"IsInternalOnly":{"description":"Indicates that this setting should be hidden to all public requests.","type":"boolean"},"Description":{"description":"Description of the App Setting","maxLength":1000,"minLength":0,"type":"string"},"ModifiedDate":{"format":"date-time","description":"Last modified date and time","type":"string"},"ModifiedBy":{"description":"Modified by","maxLength":50,"minLength":0,"type":"string"}}},"AppSettingWithOverrideDTO":{"description":"AppSetting with overridden indicator. Used in the management UI to highlight or dim entries","required":["SettingKey","TenantId","SettingGroupId","SettingSectionId","TypeId"],"type":"object","properties":{"Overridden":{"description":"Has this setting been overridden by one at a high level","type":"boolean"},"AppSettingId":{"format":"int32","description":"The setting Id","type":"integer"},"SettingKey":{"description":"Setting Key","maxLength":50,"minLength":0,"type":"string"},"TenantId":{"format":"int32","description":"Tenant Id. Zero indicates that this setting applies to all affiliates unless overridden","type":"integer"},"SettingGroupId":{"description":"The group that this setting applies to. Each client is will be interested in the settings contained in a group and it's parent groups","type":"string"},"GroupPath":{"description":"The group path","type":"string"},"SettingSectionId":{"description":"The setting UI section","type":"string"},"TypeId":{"description":"The type of setting value. e.g. HTML, TEXT,","enum":["TEXT","INTEGER","DECIMAL","BOOL","CSV","HTML","JSON","XML"],"type":"string"},"SettingValue":{"description":"The setting value.","type":"string"},"IsLocked":{"description":"Indicates that this setting can not be edited in the DCR Admin UI","type":"boolean"},"IsInternalOnly":{"description":"Indicates that this setting should be hidden to all public requests.","type":"boolean"},"Description":{"description":"Description of the App Setting","maxLength":1000,"minLength":0,"type":"string"},"ModifiedDate":{"format":"date-time","description":"Last modified date and time","type":"string"},"ModifiedBy":{"description":"Modified by","maxLength":50,"minLength":0,"type":"string"}}},"AppSettingSectionDTO":{"description":"App Setting UI Section","type":"object","properties":{"AppSettingSectionId":{"format":"int32","description":"Unique Section Id","type":"integer"},"ParentSectionId":{"format":"int32","description":"Parent Section Id","type":"integer"},"Section":{"description":"Section name","type":"string"},"Description":{"description":"Section description","type":"string"},"Ordinality":{"format":"int32","description":"The ordinality of the section","type":"integer"},"ModifiedDate":{"format":"date-time","description":"Date last modified","type":"string"},"ModifiedBy":{"description":"Last modified by","type":"string"}}},"TenantDTO":{"description":"Tenant","type":"object","properties":{"TenantId":{"format":"int32","description":"Unique Tenant Id","type":"integer"},"TenantName":{"description":"Tenant name (public display)","type":"string"},"TenantCode":{"description":"Tenant code (used for reporting and reference)","type":"string"},"TenantDescription":{"description":"Tenant Description","type":"string"},"ModifiedDate":{"format":"date-time","description":"Date last modified","type":"string"},"ModifiedBy":{"description":"Last modified by","type":"string"}}}}}