# src/dotNET/hms.entappsettings.webapi/hms.entappsettings.webapi.clientsdk/hms.entappsettings.webapi/Models/AppSettingSectionDTO.cs
// Code generated by Microsoft (R) AutoRest Code Generator 0.16.0.0
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.

namespace hms.entappsettings.webapi.clientsdk.Models
{
    using System;
    using System.Linq;
    using System.Collections.Generic;
    using Newtonsoft.Json;
    using Microsoft.Rest;
    using Microsoft.Rest.Serialization;

    /// <summary>
    /// App Setting UI Section
    /// </summary>
    public partial class AppSettingSectionDTO
    {
        /// <summary>
        /// Initializes a new instance of the AppSettingSectionDTO class.
        /// </summary>
        public AppSettingSectionDTO() { }

        /// <summary>
        /// Initializes a new instance of the AppSettingSectionDTO class.
        /// </summary>
        public AppSettingSectionDTO(int? appSettingSectionId = default(int?), int? parentSectionId = default(int?), string section = default(string), string description = default(string), int? ordinality = default(int?), DateTime? modifiedDate = default(DateTime?), string modifiedBy = default(string))
        {
            AppSettingSectionId = appSettingSectionId;
            ParentSectionId = parentSectionId;
            Section = section;
            Description = description;
            Ordinality = ordinality;
            ModifiedDate = modifiedDate;
            ModifiedBy = modifiedBy;
        }

        /// <summary>
        /// Unique Section Id
        /// </summary>
        [JsonProperty(PropertyName = "AppSettingSectionId")]
        public int? AppSettingSectionId { get; set; }

        /// <summary>
        /// Parent Section Id
        /// </summary>
        [JsonProperty(PropertyName = "ParentSectionId")]
        public int? ParentSectionId { get; set; }

        /// <summary>
        /// Section name
        /// </summary>
        [JsonProperty(PropertyName = "Section")]
        public string Section { get; set; }

        /// <summary>
        /// Section description
        /// </summary>
        [JsonProperty(PropertyName = "Description")]
        public string Description { get; set; }

        /// <summary>
        /// The ordinality of the section
        /// </summary>
        [JsonProperty(PropertyName = "Ordinality")]
        public int? Ordinality { get; set; }

        /// <summary>
        /// Date last modified
        /// </summary>
        [JsonProperty(PropertyName = "ModifiedDate")]
        public DateTime? ModifiedDate { get; set; }

        /// <summary>
        /// Last modified by
        /// </summary>
        [JsonProperty(PropertyName = "ModifiedBy")]
        public string ModifiedBy { get; set; }

    }
}